move tab state to pages level
